# internal/docs/swagger.yaml
basePath: /v1
definitions:
...
  api.ErrorResponse:
    properties:
      error:
        type: string
    type: object
...
  api.ManagerSeasonsResponse:
    properties:
      manager_id:
        type: string
      seasons:
        items:
          $ref: '#/definitions/core.ManagerSeasonRecord'
        type: array
    type: object
  api.PaginatedResponse:
    properties:
      data: {}
      page:
        type: integer
      per_page:
        type: integer
      total:
        type: integer
    type: object
...
  api.datasetCoverage:
    properties:
      from:
        type: integer
      to:
        type: integer
    type: object
  api.metaResponse:
    properties:
      coverage:
        additionalProperties:
          $ref: '#/definitions/api.datasetCoverage'
        type: object
      datasets:
        items:
          $ref: '#/definitions/core.DatasetStatus'
        type: array
      generated_at:
        type: string
      schema_hashes:
        additionalProperties:
          type: string
        type: object
      version:
        type: string
    type: object
...
  core.DatasetStatus:
    properties:
      coverage_from:
        type: integer
      coverage_to:
        type: integer
      healthy:
        type: boolean
      id:
        type: string
      last_loaded_at:
        type: string
      name:
        type: string
      required:
        type: boolean
      row_count:
        type: integer
      source:
        type: string
      tables:
        additionalProperties:
          format: int64
          type: integer
        type: object
    type: object
...
  core.LeagueConstant:
    properties:
      league:
        description: AL or NL
        type: string
      replacement_pa:
        description: replacement level runs per PA
        type: number
      runs_per_win:
        description: league-specific runs per win
        type: number
      season:
        type: integer
      total_pa:
        description: total plate appearances
        type: integer
      total_runs:
        description: total runs scored
        type: integer
      woba_avg:
        description: league average wOBA
        type: number
      wrc_per_pa:
        description: wRC per PA (excluding pitchers)
        type: number
    type: object
...
  core.Manager:
    properties:
      debut_game:
        type: string
      first_name:
        type: string
      full_name:
        type: string
      id:
        type: string
      last_game:
        type: string
      last_name:
        type: string
      player_id:
        type: string
      use_name:
        type: string
    type: object
  core.ManagerSeasonRecord:
    properties:
      g:
        type: integer
      l:
        type: integer
      manager_id:
        type: string
      rank:
        type: integer
      team_id:
        type: string
      w:
        type: integer
      year:
        type: integer
    type: object
...
  core.ParkFactorRow:
    properties:
      basic_1yr:
        description: single year (most volatile)
        type: integer
      basic_3yr:
        description: 3-year regressed
        type: integer
      basic_5yr:
        description: 5-year regressed (most stable)
        type: integer
      factor_1b:
        description: singles
        type: integer
      factor_2b:
        description: doubles
        type: integer
      factor_3b:
        description: triples
        type: integer
      factor_bb:
        description: walks
        type: integer
      factor_fb:
        description: fly balls
        type: integer
      factor_fip:
        description: FIP
        type: integer
      factor_gb:
        description: ground balls
        type: integer
      factor_hr:
        description: home runs
        type: integer
      factor_iffb:
        description: infield fly balls
        type: integer
      factor_ld:
        description: line drives
        type: integer
      factor_so:
        description: strikeouts
        type: integer
      park_id:
        type: string
      season:
        type: integer
      team_id:
        description: team playing at this park
        type: string
    type: object
...
  core.ReadinessStatus:
    properties:
      checked_at:
        type: string
      datasets:
        items:
          $ref: '#/definitions/core.DatasetStatus'
        type: array
      missing:
        items:
          type: string
        type: array
      ready:
        type: boolean
      status:
        type: string
    type: object
...
  core.WOBAConstant:
    properties:
      c_fip:
        description: FIP constant
        type: number
      r_pa:
        description: runs per plate appearance
        type: number
      r_w:
        description: runs per win
        type: number
      run_cs:
        description: caught stealing runs
        type: number
      run_sb:
        description: stolen base runs
        type: number
      season:
        type: integer
      w_1b:
        description: single
        type: number
      w_2b:
        description: double
        type: number
      w_3b:
        description: triple
        type: number
      w_bb:
        description: unintentional walk
        type: number
      w_hbp:
        description: hit by pitch
        type: number
      w_hr:
        description: home run
        type: number
      woba:
        description: league average wOBA
        type: number
      woba_scale:
        description: scaling factor
        type: number
    type: object
...
info:
  contact:
    email: info@stormlightlabs.org
    name: API Support
    url: https://github.com/stormlightlabs/baseball
  description: |
    <!-- markdownlint-disable -->

    A comprehensive REST API for baseball statistics serving data from the Lahman Baseball Database and Retrosheet

    ## Attributions

    **[SABR Lahman Database](https://sabr.org/lahman-database/)**: The information used here was obtained free of charge from and is copyrighted by Sean Lahman.

    **[Retrosheet.org](https://www.retrosheet.org/)**: The information used here was obtained free of charge from and is copyrighted by Retrosheet.

    wOBA weights and seasonal constants from **[FanGraphs Guts!](https://www.fangraphs.com/tools/guts)**, used under FanGraphs’ published terms.
  license:
    name: MPL-2.0
    url: https://opensource.org/license/mpl-2-0
  title: Baseball API
  version: "1.0"
paths:
...
  /managers:
    get:
      consumes:
      - application/json
      description: Get a paginated list of all managers
      parameters:
      - default: 1
        description: Page number
        in: query
        name: page
        type: integer
      - default: 50
        description: Results per page
        in: query
        name: per_page
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/api.PaginatedResponse'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/api.ErrorResponse'
      summary: List managers
      tags:
      - managers
  /managers/{manager_id}:
    get:
      consumes:
      - application/json
      description: Get detailed information about a specific manager including extended
        biodata from Retrosheet (debut/last game, full name, use name)
      parameters:
      - description: Manager ID (playerID)
        example: roberda07
        in: path
        name: manager_id
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/core.Manager'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/api.ErrorResponse'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/api.ErrorResponse'
      summary: Get manager by ID
      tags:
      - managers
  /managers/{manager_id}/seasons:
    get:
      consumes:
      - application/json
      description: Get all season records for a specific manager including wins, losses,
        and team rank
      parameters:
      - description: Manager ID (playerID)
        example: roberda07
        in: path
        name: manager_id
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/api.ManagerSeasonsResponse'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/api.ErrorResponse'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/api.ErrorResponse'
      summary: Get manager season records
      tags:
      - managers
  /meta:
    get:
      consumes:
      - application/json
      description: Returns API version, dataset freshness, coverage, and schema fingerprints
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/api.metaResponse'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/api.ErrorResponse'
      summary: API metadata
      tags:
      - meta
  /meta/constants/league:
    get:
      consumes:
      - application/json
      description: Returns league-specific constants for wRC+ and WAR calculations
      parameters:
      - description: Season year (returns all if omitted)
        in: query
        name: season
        type: integer
      - description: League (AL or NL)
        in: query
        name: league
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            items:
              $ref: '#/definitions/core.LeagueConstant'
            type: array
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/api.ErrorResponse'
      summary: League constants
      tags:
      - meta
  /meta/constants/park-factors:
    get:
      consumes:
      - application/json
      description: Returns FanGraphs park factors for seasons
      parameters:
      - description: Season year (returns all if omitted)
        in: query
        name: season
        type: integer
      - description: Team ID filter
        in: query
        name: team
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            items:
              $ref: '#/definitions/core.ParkFactorRow'
            type: array
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/api.ErrorResponse'
      summary: Park factors
      tags:
      - meta
  /meta/constants/woba:
    get:
      consumes:
      - application/json
      description: Returns season-specific wOBA calculation constants from FanGraphs
      parameters:
      - description: Season year (returns all if omitted)
        in: query
        name: season
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            items:
              $ref: '#/definitions/core.WOBAConstant'
            type: array
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/api.ErrorResponse'
      summary: wOBA constants
      tags:
      - meta
  /meta/datasets:
    get:
      consumes:
      - application/json
      description: Returns dataset ETL metadata and coverage
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            items:
              $ref: '#/definitions/core.DatasetStatus'
            type: array
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/api.ErrorResponse'
      summary: Dataset status
      tags:
      - meta
  /meta/readiness:
    get:
      consumes:
      - application/json
      description: Returns whether the required datasets are loaded for the core API
        routes
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/core.ReadinessStatus'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/api.ErrorResponse'
      summary: Dataset readiness
      tags:
      - meta
      - health
...
swagger: "2.0"
...